musicbox/requirements.txt
python-vlc
flask
